frontend::components

Module buttons

Standard action buttons for various controls (closing, minimising, adding a plug-in) including the associated context menu for plug-in selection.

Re-exports

pub use CloseButton_completions::Component::CloseButton;

pub use MinimiseButton_completions::Component::MinimiseButton;

pub use AddButton_completions::Component::AddButton;

Structs

AddButtonProps

Properties for the AddButton component.

CloseButtonProps

Properties for the CloseButton component.

ContextMenuProps 🔒

Properties for the ContextMenu component.

MinimiseButtonProps

Properties for the MinimiseButton component.

Functions

AddButton

A button that asynchronously fetches available plug-ins for a specific position upon click, opening a selection context menu.

CloseButton

A button utilised to close an active panel or plug-in view. The button does not close the component it is placed in per sé when clicked. Instead, the event handler provided via its onclick argument is executed.

ContextMenu 🔒

An overlay menu for keyboard- and mouse-based selection of a plug-in from a list. Currently, it is always shown centered on the screen, due to unresolved problems with mouse pointer based positioning close to the lower and right screen edge.

MinimiseButton

A button used to toggle the minimised state of a panel. The state is passed as [Signal] via the button’s only argument.