frontend::components

Module panel

This module contains the basis of the framework’s UI, the Panel. It supports dynamic splitting, resizing, and optionally minimising and closing.

Re-exports

pub use Panel_completions::Component::Panel;

Structs

ContextMenuProps 🔒

Properties for the ContextMenu component.

PanelProps

Properties for the Panel component.

Functions

ContextMenu 🔒

Local context menu for choosing the split direction (Horizontal/Vertical).

Panel

The primary visual unit of the user interface.

on_context_menu 🔒

Triggers the ContextMenu for splitting the Panel.

on_mounted

Registers the panel within the [GroupContext] post-rendering to coordinate size calculations. This is essential for panel resizing.