frontend::components::panel

Function Panel

pub fn Panel(_: PanelProps) -> Element

The primary visual unit of the user interface.

Can either be a fixed leaf ([Variant::Leaf]) displaying content, or a branch ([Variant::Branch]) that recursively splits into sub-panels.

The Panel contains state management logic for resizing and minimising, as well as for styling. Additionally, it contains information about its position on the screen.

Arguments

custom_classes - An optional String containing any number of space separated CSS classes. They are added without further validation to the class attribute of the resulting HTML element.

panel_name - The optional text that is displayed inside the PanelHeader if the panel is not headless.

headless - Whether this Panel shall have a PanelHeader containing alongside a title string controls such as a minimise button or a close button. Default value is false.

minimisable - Set to true to show a minimise button in the PanelHeader if not headless. Default value is true.

required - Set to false to show a close button in the PanelHeader if not headless. Default value is true.

min_size - Set this value to something above 0, if you want to restrict resizing this Panel to at least min_size pixels along the resizable axis. Does not affect the other axis when the containing PanelGroup is resized. Default value is 0. Negative values, while none-sense, are acceepted and may lead to undefined behaviour.

position - One of the four [Position] variants indicating where the Panel is positioned on the screen.

children - Any Dioxus component that shall be displayed inside the Panel.

Props

For details, see the props struct definition.

custom_classes : String

panel_name : String

headless : bool

minimisable : bool

required : bool

min_size : i32

position : Position

children : Element