frontend::components::panel

Function on_mounted

pub async fn on_mounted(
    event: MountedEvent,
    context: Option<GroupContext>,
    uuid: Uuid,
    min_size: i32,
)

Registers the panel within the [GroupContext] post-rendering to coordinate size calculations. This is essential for panel resizing.

Arguments

event - The mounted event triggered when the corresponding Panel is rendered. Contains the required information about the Panel’s dimensions on the screen.

context - The [GroupContext] that is available to the Panel registering via this function. If the Option is None, the function returns immediately.

uuid - A Uuid value assigned to each element inside a [GroupContext] to track them throughout layout changes.

min_size - The minimum display size of the registering Panel.