frontend::components::panel_header

Function PanelHeader

pub fn PanelHeader(_: PanelHeaderProps) -> Element

The header component of a [Panel]. Also used by the [TabbedGroup] as tab header.

Renders the panel’s name left-aligned and passes through a collection of control buttons (e.g. close or minimise actions) to the right side. The buttons are displayed from left to right in the order they are handed over.

Props

For details, see the props struct definition.

panel_name : String

class : Option<String>

buttons : Element

onclick : EventHandler<MouseEvent>