frontend::components::plugin_panel

Function PluginPanel

pub fn PluginPanel(_: PluginPanelProps) -> Element

A Panel dedicated to hosting and rendering a specific plug-in.

Encapsulates the plug-in inside an HTML iframe enforced with custom sandbox restrictions, managing token-authorised message-passing to the back-end.

Arguments

Most of this panel’s arguments/props are equal in meaning to those of the default Panel. Hence, we only cover the additional external_plugin argument here.

external_plugin - Set to a [Manifest][plugin::Manifest] value to provide the plug-in panel with a pre-set plug-in. For reference, this is used in the [TabbedGroup] component.

Props

For details, see the props struct definition.

panel_name : String

headless : bool

position : Position

min_size : i32

required : bool

external_plugin : Option<plugin::Manifest>

children : Element