frontend::components::tabbed_group

Function TabbedGroup

pub fn TabbedGroup(_: TabbedGroupProps) -> Element

This function generates an [Element] that groups [PluginPanels] in a tabbed view.

Automatically organises launched plug-ins into clickable tabs, orchestrates tab removal, focus shifting, and manages custom fallback placeholder content when no tabs are active.

Can be used as resizable component in a [PanelGroup].

Arguments

position - The [Position] variant determines some behavioural traits with respect to rendering.

min_size - The minimum size this element may be shrinked to.

children - Child elements to display in the empty panel body as placeholder text.

Props

For details, see the props struct definition.

position : Position

min_size : i32

children : Element