frontend

Module slow_query_recorder

Structs

SlowQueryEvent 🔒

SlowQueryEventHandler 🔒

SlowQueryRecorder

SlowQueryRecorder is responsible for recording slow queries.

SlowQueryTimer

SlowQueryTimer is used to log slow query when it’s dropped. In drop(), it will check if the query is slow and send the slow query event to the handler.

Constants

DEFAULT_SLOW_QUERY_EVENTS_CHANNEL_SIZE 🔒

DEFAULT_SLOW_QUERY_TABLE_TTL 🔒

SLOW_QUERY_TABLE_COST_COLUMN_NAME 🔒

SLOW_QUERY_TABLE_IS_PROMQL_COLUMN_NAME 🔒

SLOW_QUERY_TABLE_NAME 🔒

SLOW_QUERY_TABLE_PROMQL_END_COLUMN_NAME 🔒

SLOW_QUERY_TABLE_PROMQL_RANGE_COLUMN_NAME 🔒

SLOW_QUERY_TABLE_PROMQL_START_COLUMN_NAME 🔒

SLOW_QUERY_TABLE_PROMQL_STEP_COLUMN_NAME 🔒

SLOW_QUERY_TABLE_QUERY_COLUMN_NAME 🔒

SLOW_QUERY_TABLE_THRESHOLD_COLUMN_NAME 🔒

SLOW_QUERY_TABLE_TIMESTAMP_COLUMN_NAME 🔒