Crate frontend

Modules

constants

key_press_handling

player

utils

Structs

Materials

Position
The internal position of something in the world; where in the world it can be considered to be.

Tile

Functions

camera_player_tracking
makes the camera follow the player. TODO: when the player is in corners, make the camera stay in the same spot. NOTE: this will first require fixing how the tiles are drawn

main

player_movement_input_handling
Moves the player by changing it’s internal dungeon::Point according to the input.

setup

smooth_player_movement
Moves the player between squares smoothly.

spawn_player_and_board