frontend::instance
Module prom_store

Structs

ExportMetricHandler

This handler is mainly used for frontend or standalone to directly import the metrics collected by itself, thereby avoiding importing metrics through the network, thus reducing compression and network transmission overhead, so only implement PromStoreProtocolHandler::write method.

Constants

SAMPLES_RESPONSE_TYPE 🔒

Functions

is_supported 🔒

negotiate_response_type 🔒

Negotiating the content type of the remote read response.

to_query_result 🔒